+++ Engine/physics/integrators/KDKScheme.h
@@ -8,8 +8,8 @@ class SimBox;
 
 class KDKScheme {
 public:
-    void pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, double dt) const;
+    void pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, float dt) const;
 
-    static void halfKick(AtomStorage& atomStorage, std::size_t atomIndex, double dt);
-    static void drift(AtomStorage& atomStorage, std::size_t atomIndex, double dt);
+    static void halfKick(AtomStorage& atomStorage, std::size_t atomIndex, float dt);
+    static void drift(AtomStorage& atomStorage, std::size_t atomIndex, float dt);
 };
