+++ Engine/physics/integrators/LangevinScheme.cpp
@@ -2,6 +2,6 @@
 
 #include "VerletScheme.h"
 
-void LangevinScheme::pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, double dt) const {
+void LangevinScheme::pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, float dt) const {
     VerletScheme{}.pipeline(atomStorage, box, forceField, dt);
 }
