+++ Engine/physics/integrators/RK4Scheme.cpp
@@ -2,6 +2,6 @@
 
 #include "VerletScheme.h"
 
-void RK4Scheme::pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, double dt) const {
+void RK4Scheme::pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, float dt) const {
     VerletScheme{}.pipeline(atomStorage, box, forceField, dt);
 }
