+++ Engine/physics/integrators/VerletScheme.h
@@ -8,8 +8,8 @@ class SimBox;
 
 class VerletScheme {
 public:
-    void pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, double dt) const;
+    void pipeline(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, float dt) const;
 
-    static void predict(AtomStorage& atomStorage, std::size_t atomIndex, double dt);
-    static void correct(AtomStorage& atomStorage, std::size_t atomIndex, double dt);
+    static void predict(AtomStorage& atomStorage, std::size_t atomIndex, float dt);
+    static void correct(AtomStorage& atomStorage, std::size_t atomIndex, float dt);
 };
